Get initial credentials also from secrets
Secrets are supposed to be mounted in subdirectories of /run/secrets/pgusers/.
They should have `username` and `password` keys.

Limitations: It does not check for duplicate user names or overlap with
(legacy) environment variables setting credentials.

Regarding startup works similar to the previous environment variable based logic.
It will only start if one or both of the following is true.

 * `/run/secrets/pguser/user` are valid credentials and POSTGRESQL_DATABASE is set
   (a database will be created).

 * `/run/secrets/pguser/admin` are valid credentials and the username is `postgres`.

`/run/secrets/pguser/master` is supposed to be the replication master user, if mounted.