tests-openshift: race condition in getting pod IP
After successful pod scale to 2, it takes some time to have the
replica ready, and it probably takes some time to get allocated
the IP address.  So rather wait for the pod to be ready before we
expect that the IP is available.

We could in future make the 'ct_os_get_pod_ip' accept some option
which would wait till the IP is available.