Let the username/password validation on PostgreSQL
Remove the artificial limitations for the username/password format
given by our scripting, and start using 'psql --set' feature to
avoid sql injection (previously we guarded against sql injection
by the regexps).  Benefits sum-up:

 - protection against sql injection for free
 - our passwords can contain spaces and other special characters,
   using them in passwords is generally considered to be good
   practice
 - simplified, more readable "usage" output

The downside is that the rejection of the wrongly specified
passwords/usernames is a bit delayed (till the PostgreSQL is is
running to perform the actual validation).  But because PostgreSQL
itself doesn't create any unexpected requirements on the values,
it is newly more unlikely to get such failure.