update imagestreams to registry.redhat.io
registry.redhat.io is the future of the redhat container catalog, registry.access.redhat.com is going to be shut down at some point in the future. The new registry requires authentication and the local reference policy allows the openshift internal registry to perform that authentication on the user's behalf so that every use on a cluster doesn't have to provide their own registry.redhat.io credentials.