exclude container user's id when setting up passwd file
When a container is run with a recent enough podman version and using the --user
cli argument the /etc/passwd file inside the container includes a line with
the user defined via --user.

$ podman run --rm -ti --user 1001 rhel7 bash -c "cat /etc/passwd | grep 1001"
  1001:x:1001:0:container user:/:/bin/sh

Since postgresql creates a postgres user with the user id of the container user
we need to exclude the passwd line containing said user from the generated
passwd file else the initial database will not be owned by the postgres user.