Demonstrate problem in running tests
This is caused by the change in running tests, when we wanted to run all tests by default.
However, when doing so, `set -e` setting does not work, because the -e is ignored in all
commands under if statement. This is the relevant documentation in the man page bash(1).

"""
If a compound command or shell function executes in a context where -e is being ignored,
none of the commands executed within the compound command or function body will be affected
by the -e setting, even if -e is set and a command returns a failure status.

If a compound command or shell function sets -e while executing in a context where -e is
ignored, that setting will not have any effect until the compound command or the command
containing the function call completes.
"""