Remove FAIL_QUICKLY feature because it hides failures in the test functions
There is a problem in the way test cases could be run all, even if they failed. Running the test function inside an IF statment disables 'set -e' in the function and cannot be re-set. The problem is demonstrated in https://github.com/sclorg/postgresql-container/pull/363 which added 'false' call in the beginning of the test functions, but the test suite still passed.