Use the podman workaround for exec and redirection in migration test
For the migration test, the change is more complicated because there is no
easy way to direct longer text into the psql input from file, when redirecting
does not work with exec and the container itself does not have the file.

It was solved by running a separate container, because with podman run
the problem with redirection does not occur.