Revert "each container in run_change_password_test mounts its own volume"
This reverts commit 62a0e88ae0e5f19c3dad1cb6fea3778dfa2dbf2f.

Creating a new volume directory was actually NOOP, because the variable
volume_options was not updated. What we need to get rid of SELinux
messages that this fix tried to address is to properly shut down
previous container. That will be done in the following commit.

Background:
podman's :Z modificator for volumes works the way that the latest
container run with the same volume directory mounted with :Z modificator
has access, the previous containers are kept running, but access to the
shared directory is suddenly removed. That caused the first instance of
PostgreSQL server to crash with SIGSEGV actually, while triggering some
SELinux error message during that.