More shared functions and test log readability
This combines two changes that both improve test script, so I did not
bother to split it to two commits. One is using more functions from the
test-lib shared library, and second is removing expected warnings from
the test log, to make the log more readable and easier to spot the
issues.