Fix pagila RPM url and wait patiently for the server start
Previous epel7 mirror URL does not work after RHEL7 EOL, but the same
RPM is still available in koji. A test for a correct download should
hopefully make the download failure easier to debug in the future.

The check functions did not actually wait when the first podman exec
failed, so in case the server was not up soon, the check failed too
early. It is ok to wait for the server a bit, so let's repeat the
podman exec with psql until it succeeds or 30s pass.