Fix error in `run_pgaudit_test()`
Initially command `test was missing`, but after it was added
test failed and crashed in a way that no more tests were run.
This behavior is known issue tracked here in issue #537

Problem is that `DOCKER_EXTRA_ARGS` variable is set already
which it shouldn't be, so the `test` command returns 0,
so no `cp` happens and we get the crash.

Variable is empty, but apparently do exist as
`test -v DOCKER_EXTRA_ARGS` is `TRUE`.

Introduced by PR #399